gobiiproject/gobii-client/src/main/java/org/gobiiproject/gobiiclient/dtorequests/DtoRequestReference.java
// ************************************************************************
// (c) 2016 GOBii Project
// Initial Version: Phil Glaser
// Create Date:   2016-03-25
// ************************************************************************
package org.gobiiproject.gobiiclient.dtorequests;

import org.gobiiproject.gobiiclient.core.DtoRequestProcessor;
import org.gobiiproject.gobiiclient.core.TypedRestRequest;
import org.gobiiproject.gobiiclient.core.Urls;
import org.gobiiproject.gobiimodel.dto.container.DisplayDTO;
import org.gobiiproject.gobiimodel.dto.container.ReferenceDTO;
import org.gobiiproject.gobiimodel.dto.types.ControllerType;
import org.gobiiproject.gobiimodel.dto.types.ServiceRequestId;
import org.gobiiproject.gobiimodel.types.SystemUserDetail;
import org.gobiiproject.gobiimodel.types.SystemUserNames;
import org.gobiiproject.gobiimodel.types.SystemUsers;

public class DtoRequestReference {

    public ReferenceDTO process(ReferenceDTO referenceDTO) throws Exception {

        return new DtoRequestProcessor<ReferenceDTO>().process(referenceDTO,
                ReferenceDTO.class,
                ControllerType.LOADER,
                ServiceRequestId.URL_REFERENCE);

    } // getPing()

}
